Add setting methods of all class type
