+++ Sources/Keychain.swift
@@ -54,3 +54,41 @@ extension Keychain {
         return .init()
     }
 }
+
+// MARK: - KeychainItemClassType
+
+public protocol KeychainItemClassType: KeychainCommonItemAttributes {
+    
+    var rawClassValue: KeychainItemClassValue { get }
+    
+    func makeBasicQuery() -> KeychainBasicQuerySetter<Self>
+    func makeSaveQuery() -> KeychainSaveQuerySetter<Self>
+    func makeSearchQuery() -> KeychainSearchQuerySetter<Self>
+    func makeUpdateQuery() -> KeychainUpdateQuerySetter<Self>
+    func makeDeleteQuery() -> KeychainDeleteQuerySetter<Self>
+}
+
+extension KeychainItemClassType {
+    
+    // MARK: Default implementations
+    
+    public func makeBasicQuery() -> KeychainBasicQuerySetter<Self> {
+        return .init(classValue: rawClassValue)
+    }
+    
+    public func makeSaveQuery() -> KeychainSaveQuerySetter<Self> {
+        return .init(classValue: rawClassValue)
+    }
+    
+    public func makeSearchQuery() -> KeychainSearchQuerySetter<Self> {
+        return .init(classValue: rawClassValue)
+    }
+    
+    public func makeUpdateQuery() -> KeychainUpdateQuerySetter<Self> {
+        return .init(classValue: rawClassValue)
+    }
+    
+    public func makeDeleteQuery() -> KeychainDeleteQuerySetter<Self> {
+        return .init(classValue: rawClassValue)
+    }
+}
